Implement Azure Maps.

+++ src/UnitTestSample/Models/WeatherForecast.cs
@@ -6,9 +6,9 @@ namespace UnitTestSample.Models
     {
         public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+        public double TemperatureC { get; set; }
 
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+        public double TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
         public string Summary { get; set; }
     }
